packages/ingest/src/sidecar/scripts/requirements.txt
faster-whisper>=1.0.0
pyannote.audio>=3.1.0
scenedetect[opencv]>=0.6.4
Pillow>=10.0.0
imagehash>=4.3.1
